% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_slots.R
\name{check_slots}
\alias{check_slots}
\title{Checks slots with yaml rules}
\usage{
check_slots(
  scenarios,
  yaml_rule_files,
  scenario_dir,
  output_dir,
  yaml_dir,
  out_fl_nm = "verification_output"
)
}
\arguments{
\item{scenarios}{A vector of folder names inside the scenario_dir,
which stores the scenario output.}

\item{yaml_rule_files}{A vector of yaml files.}

\item{scenario_dir}{Directory where scenarios are stored.}

\item{output_dir}{Directory where summary and log files are saved.}

\item{yaml_dir}{Directory where yaml files are stored.}

\item{out_fl_nm}{The name of the output files, default to
verification_output.}
}
\value{
Writes pass and fails to summary verification file and outputs
  details to log_file.txt.
}
\description{
Add a description of the function here.
}
\examples{
scenarios <- c("MRM_Avg,ModelBase,RulesBase,Run-2019-10",
  "MRM_Avg,ModelBase,RulesBase,Run-2019-11")
yaml_rule_files <- c("check_lb_res.yaml", "check_ub_outflow.yaml")
scenario_dir <- "C:/User/Project/Scenario/"
output_dir <- "C:/User/Project/ScenarioSet/allScenarios/basicChecks"
yaml_dir <- "C:/User/Project/Code/"
check_slots(scenarios, yaml_rule_files, scenario_dir, output_dir, yaml_dir)

}
