% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_error.R
\name{dir_error}
\alias{dir_error}
\title{Check that directory exists}
\usage{
dir_error(dir_check)
}
\arguments{
\item{dir_check}{Directory.}
}
\value{
Error if directory does not exist.
}
\description{
Checks if directory exists and errors if it doesnt.
}
\examples{
dir_error("C:/User/Project/")
}
