% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scenario.R
\name{read_scenario}
\alias{read_scenario}
\title{Read and append all data_files based on file type}
\usage{
read_scenario(data_files)
}
\arguments{
\item{data_files}{A vector of data files produced by get_datafiles().}
}
\value{
A dataframe containg RW output organized into
}
\description{
The function reads RiverWare data files based on file type.
}
\examples{
read_scenario(c('ReservoirOutput.csv'))

dontshow{
read_scenario('ReservoirOutput.csv')
}

dontrun{
read_scenario(c('badtype.txt'))
}

}
