% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datafiles.R
\name{get_datafiles}
\alias{get_datafiles}
\title{Get all data files used in yaml rules}
\usage{
get_datafiles(yaml_files)
}
\arguments{
\item{yaml_files}{A vector of yaml files.}
}
\value{
A vector of data files.
}
\description{
Add a description of the function here.
}
\examples{
get_datafiles(c('check_lb_res.yaml', 'check_ub_outflow.yaml'))

dontrun{
get_datafiles(c('badyaml.txt'))
}
}
