% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scenario.R
\name{read_scenario}
\alias{read_scenario}
\title{Read and append all data_files based on file type}
\usage{
read_scenario(data_files)
}
\arguments{
\item{data_files}{A vector of data files produced by get_datafiles().}
}
\value{
A dataframe containg RW output with columns as slots and rows
  with timestep and trace number combination.
}
\description{
The function reads RiverWare output based on file type and binds
data together if there are multiple \code{data_files}.
}
\examples{
read_scenario(c('ReservoirOutput.csv', UBRes.rdf))

dontrun{
read_scenario(c('badtype.txt'))
}

}
